requirements.txt
asgiref==3.6.0
click==8.1.3
Django==3.2
Faker==16.6.0
flake8==6.0.0
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
python-dotenv==0.21.1
pytest==7.2.1
django-debug-toolbar==3.4.0
django-split-settings==1.2.0
parameterized==0.8.1